Small bugs fixed, created helpers folder
