+++ package.json
@@ -14,6 +14,8 @@
     "crypto-js": "^4.0.0",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
+    "express-fileupload": "^1.2.1",
+    "googleapis": "^67.1.0",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^5.11.15",
     "nodemailer": "^6.4.17",
